Add menu command to Exposify prototype and setup checking for gradebooks
I may want to create menus dynamically in the future, so I create the
menu object on the Exposify constructor now. I also added a few
functions to get, set, and check whether a gradebook has been set up on
a given sheet, so certain commands are not available until this has
been done.